requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
--find-links https://download.pytorch.org/whl/torch_stable.html
--find-links https://storage.googleapis.com/jax-releases/jax_releases.html

absl-py==1.0.0
    # via tensorboard
asttokens==2.0.5
    # via stack-data
autoflake==1.4
    # via -r requirements.in
backcall==0.2.0
    # via ipython
black==21.12b0
    # via
    #   -r requirements.in
    #   ipython
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   black
    #   pip-tools
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via ipython
executing==0.8.2
    # via stack-data
fonttools==4.29.0
    # via matplotlib
google-auth==2.5.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
grpcio==1.43.0
    # via tensorboard
idna==3.3
    # via requests
importlib-metadata==4.10.1
    # via markdown
ipython==8.0.1
    # via -r requirements.in
isort==5.10.1
    # via -r requirements.in
jedi==0.18.1
    # via ipython
json5==0.9.6
    # via -r requirements.in
kiwisolver==1.3.2
    # via matplotlib
kornia==0.6.3
    # via -r requirements.in
markdown==3.3.6
    # via tensorboard
matplotlib==3.5.1
    # via -r requirements.in
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
numpy==1.22.1
    # via
    #   -r requirements.in
    #   matplotlib
    #   opencv-python
    #   tensorboard
    #   torchvision
oauthlib==3.2.0
    # via requests-oauthlib
opencv-python==4.5.5.62
    # via -r requirements.in
packaging==21.3
    # via
    #   kornia
    #   matplotlib
paho-mqtt==1.6.1
    # via -r requirements.in
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.0.0
    # via
    #   matplotlib
    #   torchvision
pip-tools==6.4.0
    # via -r requirements.in
platformdirs==2.4.1
    # via black
prompt-toolkit==3.0.24
    # via ipython
protobuf==3.19.4
    # via tensorboard
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyflakes==2.4.0
    # via autoflake
pygame==2.1.2
    # via -r requirements.in
pygments==2.11.2
    # via ipython
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
requests==2.27.1
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   absl-py
    #   asttokens
    #   google-auth
    #   grpcio
    #   python-dateutil
stack-data==0.1.4
    # via ipython
tensorboard==2.8.0
    # via -r requirements.in
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
timm==0.5.4
    # via -r requirements.in
tomli==1.2.3
    # via
    #   black
    #   pep517
torch==1.10.1+cu111
    # via
    #   -c constraints.in
    #   -r requirements.in
    #   kornia
    #   timm
    #   torchvision
torchsummary==1.5.1
    # via -r requirements.in
torchvision==0.11.2+cu111
    # via
    #   -c constraints.in
    #   -r requirements.in
    #   timm
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.0.1
    # via
    #   black
    #   torch
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.2
    # via tensorboard
wheel==0.37.1
    # via
    #   pip-tools
    #   tensorboard
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
